cni: cancel pod sandbox add requests if the pod's UID or MAC changes
If the the pod's UID changes that means the pod was deleted
and re-created. There's no point in continuing this sandbox
request as kubelet will just be tearing it down soon anyway.

If the pod's MAC changes, that means the master was behind and
set the IPAM annotation on a new instance of the pod (since
the master just uses Patch with namespace+name and ignores
UID), and this sandbox will be torn down soon as well so
kubelet can start the newer one.

Signed-off-by: Dan Williams <dcbw@redhat.com>