cni: prefer pod UID from runtime rather than apiserver
Use a pod UID from the runtime to close the race between when kubelet
starts the sandbox request and when cni.go gets the pod from the
informer cache or the apiserver. If the pod was deleted and
recreated during that window the sandbox could be configured for
the new pod instance, only to be torn down soon and possibly
confuse the new sandbox ADD for the new pod instance.

Signed-off-by: Dan Williams <dcbw@redhat.com>